requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.3.4
pydantic==2.5.3
pydantic-settings==2.1.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.1
httpx==0.25.2